requirements.txt
discord.py==2.4.0
aiohttp==3.9.5
google-generativeai==0.8.3
